+++ py/src/braintrust/integrations/agno/test_agno.py
@@ -11,9 +11,9 @@ from braintrust import logger
 from braintrust.integrations.agno import setup_agno
 from braintrust.integrations.agno import tracing as agno_tracing_module
 from braintrust.integrations.agno.patchers import wrap_agent, wrap_team
+from braintrust.integrations.test_utils import verify_autoinstrument_script
 from braintrust.logger import start_span
 from braintrust.test_helpers import init_test_logger
-from braintrust.wrappers.test_utils import verify_autoinstrument_script
 
 from ._test_agno_helpers import (
     PROJECT_NAME,
