+++ py/src/braintrust/integrations/anthropic/cassettes/test_auto_anthropic.yaml
