+++ py/src/braintrust/integrations/auto_test_scripts/test_auto_litellm.py
@@ -1,14 +1,10 @@
 """Test auto_instrument for LiteLLM."""
 
-from pathlib import Path
-
 import litellm
 from braintrust.auto import auto_instrument
 from braintrust.integrations.litellm import LiteLLMIntegration
-from braintrust.wrappers.test_utils import autoinstrument_test_context
+from braintrust.integrations.test_utils import autoinstrument_test_context
 
-
-_CASSETTES_DIR = Path(__file__).resolve().parent.parent / "litellm" / "cassettes"
 
 # 1. Verify not patched initially
 assert not LiteLLMIntegration.patchers[0].is_patched(litellm, None)
@@ -26,7 +22,7 @@ results2 = auto_instrument(openai=False)
 assert results2.get("litellm") == True
 
 # 4. Make API call and verify span
-with autoinstrument_test_context("test_auto_litellm", cassettes_dir=_CASSETTES_DIR) as memory_logger:
+with autoinstrument_test_context("test_auto_litellm", integration="litellm") as memory_logger:
     response = litellm.completion(
         model="gpt-4o-mini",
         messages=[{"role": "user", "content": "Say hi"}],
