+++ py/src/braintrust/integrations/auto_test_scripts/test_auto_mistral.py
@@ -1,10 +1,9 @@
 """Test auto_instrument for Mistral."""
 
 import os
-from pathlib import Path
 
 from braintrust.auto import auto_instrument
-from braintrust.wrappers.test_utils import autoinstrument_test_context
+from braintrust.integrations.test_utils import autoinstrument_test_context
 
 
 try:
@@ -19,9 +18,8 @@ assert results.get("mistral") == True
 results2 = auto_instrument()
 assert results2.get("mistral") == True
 
-MISTRAL_CASSETTES_DIR = Path(__file__).resolve().parent.parent / "mistral" / "cassettes"
 
-with autoinstrument_test_context("test_auto_mistral", cassettes_dir=MISTRAL_CASSETTES_DIR) as memory_logger:
+with autoinstrument_test_context("test_auto_mistral", integration="mistral") as memory_logger:
     client = Mistral(api_key=os.environ.get("MISTRAL_API_KEY"))
     response = client.chat.complete(
         model="mistral-small-latest",
