+++ py/src/braintrust/integrations/auto_test_scripts/test_auto_openai.py
@@ -1,15 +1,11 @@
 """Test auto_instrument for OpenAI."""
 
 import inspect
-from pathlib import Path
 
 import openai
 from braintrust.auto import auto_instrument
-from braintrust.wrappers.test_utils import autoinstrument_test_context
+from braintrust.integrations.test_utils import autoinstrument_test_context
 from wrapt import FunctionWrapper
-
-
-_CASSETTES_DIR = Path(__file__).resolve().parent.parent / "openai" / "cassettes"
 
 
 def _is_braintrust_wrapped() -> bool:
@@ -30,7 +26,7 @@ results2 = auto_instrument()
 assert results2.get("openai") == True
 
 # 4. Make API call and verify span
-with autoinstrument_test_context("test_auto_openai", cassettes_dir=_CASSETTES_DIR) as memory_logger:
+with autoinstrument_test_context("test_auto_openai", integration="openai") as memory_logger:
     client = openai.OpenAI()
     response = client.chat.completions.create(
         model="gpt-4o-mini",
