+++ py/src/braintrust/integrations/auto_test_scripts/test_auto_openai_agents.py
@@ -1,15 +1,13 @@
 """Test auto_instrument for the OpenAI Agents SDK."""
 
 import asyncio
-from pathlib import Path
 
 import agents
 from braintrust.auto import auto_instrument
 from braintrust.integrations.openai_agents import BraintrustTracingProcessor
-from braintrust.wrappers.test_utils import autoinstrument_test_context
+from braintrust.integrations.test_utils import autoinstrument_test_context
 
 
-_CASSETTES_DIR = Path(__file__).resolve().parent.parent / "openai_agents" / "cassettes"
 TEST_MODEL = "gpt-4o-mini"
 TEST_PROMPT = "What is 2+2? Just the number."
 TEST_AGENT_INSTRUCTIONS = "You are a helpful assistant. Be very concise."
@@ -29,7 +27,7 @@ results2 = auto_instrument()
 assert results2.get("openai_agents") == True
 assert _has_braintrust_processor()
 
-with autoinstrument_test_context("test_auto_openai_agents", cassettes_dir=_CASSETTES_DIR) as memory_logger:
+with autoinstrument_test_context("test_auto_openai_agents", integration="openai_agents") as memory_logger:
     from agents import Agent
     from agents.run import AgentRunner
