+++ py/src/braintrust/integrations/auto_test_scripts/test_patch_litellm_aresponses.py
@@ -1,20 +1,17 @@
 """Test that patch_litellm() patches aresponses."""
 
 import asyncio
-from pathlib import Path
 
 import litellm
 from braintrust.integrations.litellm import patch_litellm
-from braintrust.wrappers.test_utils import autoinstrument_test_context
+from braintrust.integrations.test_utils import autoinstrument_test_context
 
-
-_CASSETTES_DIR = Path(__file__).resolve().parent.parent / "litellm" / "cassettes"
 
 patch_litellm()
 
 
 async def main():
-    with autoinstrument_test_context("test_patch_litellm_aresponses", cassettes_dir=_CASSETTES_DIR) as memory_logger:
+    with autoinstrument_test_context("test_patch_litellm_aresponses", integration="litellm") as memory_logger:
         response = await litellm.aresponses(
             model="gpt-4o-mini",
             input="What's 12 + 12?",
