+++ py/src/braintrust/integrations/litellm/test_litellm.py
@@ -6,8 +6,8 @@ import litellm
 import pytest
 from braintrust import Attachment, logger
 from braintrust.integrations.litellm import patch_litellm
+from braintrust.integrations.test_utils import assert_metrics_are_valid, verify_autoinstrument_script
 from braintrust.test_helpers import assert_dict_matches, init_test_logger
-from braintrust.wrappers.test_utils import assert_metrics_are_valid, verify_autoinstrument_script
 
 
 TEST_ORG_ID = "test-org-litellm-py-tracing"
