+++ py/src/braintrust/integrations/mistral/test_mistral.py
@@ -15,8 +15,8 @@ from braintrust.integrations.mistral.tracing import (
     _chat_complete_wrapper,
     sanitize_mistral_logged_value,
 )
+from braintrust.integrations.test_utils import assert_metrics_are_valid, verify_autoinstrument_script
 from braintrust.test_helpers import init_test_logger
-from braintrust.wrappers.test_utils import assert_metrics_are_valid, verify_autoinstrument_script
 
 
 pytest.importorskip("mistralai")
