+++ py/src/braintrust/integrations/openai/test_oai_attachments.py
@@ -5,8 +5,8 @@ import time
 import openai
 import pytest
 from braintrust import Attachment, logger, wrap_openai
+from braintrust.integrations.test_utils import assert_metrics_are_valid
 from braintrust.test_helpers import init_test_logger
-from braintrust.wrappers.test_utils import assert_metrics_are_valid
 
 
 PROJECT_NAME = "test-project-openai-attachment-processing"
