+++ py/src/braintrust/integrations/openai/test_openai_openrouter_gateway.py
@@ -11,8 +11,8 @@ import time
 
 import pytest
 from braintrust import logger, wrap_openai
+from braintrust.integrations.test_utils import assert_metrics_are_valid
 from braintrust.test_helpers import init_test_logger
-from braintrust.wrappers.test_utils import assert_metrics_are_valid
 from openai import AsyncOpenAI, OpenAI
