--- py/src/braintrust/wrappers/cassettes/test_agent_with_model_settings_override_in_input.yaml
@@ -1,116 +1,0 @@
-interactions:
-- request:
-    body: '{"messages":[{"role":"user","content":"Tell me a story"}],"model":"gpt-4o-mini","max_completion_tokens":200,"stream":false,"temperature":0.9}'
-    headers:
-      Accept:
-      - application/json
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '141'
-      Content-Type:
-      - application/json
-      Cookie:
-      - __cf_bm=SnTjRxRlQXUGj3jo4FeKDaKGiTqzcsdhS_LPeF2hwmg-1766265191-1.0.1.1-aN3W.WEa6IjC4znMp0fTYaZOZeW.3xU8ex4fRGFibZ0fuR86Piz9hDDAEvh8Z9ewQ9WHoKwdnsDvPE5KIpoZY4xurXV7VS4TcwfukWZofIk;
-        _cfuvid=tXOZ7vGE2DBF6L6fDg_veKtSaUVC4UPotJDezWYoYXI-1766265191281-0.0.1.1-604800000
-      Host:
-      - api.openai.com
-      User-Agent:
-      - pydantic-ai/1.37.0
-      X-Stainless-Arch:
-      - arm64
-      X-Stainless-Async:
-      - async:asyncio
-      X-Stainless-Lang:
-      - python
-      X-Stainless-OS:
-      - MacOS
-      X-Stainless-Package-Version:
-      - 2.14.0
-      X-Stainless-Runtime:
-      - CPython
-      X-Stainless-Runtime-Version:
-      - 3.13.3
-      x-stainless-read-timeout:
-      - '600'
-      x-stainless-retry-count:
-      - '0'
-    method: POST
-    uri: https://api.openai.com/v1/chat/completions
-  response:
-    body:
-      string: !!binary |
-        H4sIAAAAAAAAAwAAAP//jFXbbttGEH3XVxzwWRIk1bVTvxZFAzRFgKZo6zaBMNodkmMtd5ndoRQi
-        MJAPSX8uX1LsUrbkXoC+CBTPXM45Mxp9nAGV2OoWlWlJTde7xbdhvPvJ1rT6+ft+/cPu11fDvfnt
-        93791c3d3bGa54ywu2ejj1lLE7resUrwE2wik3Kuur65vt5cf71Z3RSgC5ZdTmt6XVyFRSdeFpvV
-        5mqxulmsX5yy2yCGU3WLP2YA8LF8Zp7e8ofqFqv545uOU6KGq9unIKCKweU3FaUkSclrNT+DJnhl
-        X6i/9oYx9MGDoNIxJD+9H0i8womqYxzEOWoYnpM6ttixHpk93JBaNDE/tuJcAnkLQuop7p34BlEO
-        HOfQliPDyYEtyCM4C+OC2Xe05whPHVt854TSEi8l4RjiPrWhx5FSZhWZ0hAZGsOBEWqomH0uX4ok
-        aEsKsiF6trkXjuRcmoPJtDiKthBNCEePwcv7gdFzTMGTEx2XeBM6hmklk+jYBSthSG6c49iKY4RM
-        PsE4MftC32YgRrZz7AYFOQfNinLjnsokJo6Tl9qyRKSejZDDkcblW//WF7VF395nYjXFqbRYhg+K
-        +yEp6hDRSkLai3O5mIlU61n5xCBH5eZJQxROyAZko8RrmDwomyjBL/GSsWMnfChOkYIPHMepHFoq
-        09MQR2iAsnPzwmoKyq6Xr1qC2eXojrsQxyV+mVYk5ukNziLUyj7X6rEbJxF5olla8K7U3w0j6NQ6
-        69CA1FKec7HMRqZuWin+0HMU9obTHG3oswOF/cnG0tFQr8OU3ZWpT+ajYYopTySTeHQilSm89gxL
-        4xyEMQy+QSPRnfbxlTjCQZLo41KdtnKJN2XFEnZRmlYXPJ72wgwx787kWcsRjoamVY6w5A1bONkz
-        0uBdzjstB0jiEl8+ff4xLic98y+f/kTKAyWx8wwZ8png5S84cj0kylfED85dAOR90Eljvh3vTsjD
-        07Vwoelj2KW/pVa1eEntNv/Ygs+XwbFvtK0K/jAD3pW7NDw7NVUfQ9frVsOeS8P1eipYna/hGdys
-        Tker0qDkLoCntGf1tpaVxKWLy1YZMi3bc+r5DNJgJVwAswvd/6Tzb7Un7eKb/1P+DBjDvbLd9pGt
-        mOeSz2GR87/Ff4U9uVwIV4njQQxvVTjmWViuaXDTDa/SmJS7bS2+4dhHmQ553W+Zr78xG66vXlSz
-        h9lfAAAA//8DAHvMFjzWBgAA
-    headers:
-      CF-RAY:
-      - 9b1224c73cd50fb5-LAX
-      Connection:
-      - keep-alive
-      Content-Encoding:
-      - gzip
-      Content-Type:
-      - application/json
-      Date:
-      - Sat, 20 Dec 2025 21:13:30 GMT
-      Server:
-      - cloudflare
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      Transfer-Encoding:
-      - chunked
-      X-Content-Type-Options:
-      - nosniff
-      access-control-expose-headers:
-      - X-Request-ID
-      alt-svc:
-      - h3=":443"; ma=86400
-      cf-cache-status:
-      - DYNAMIC
-      openai-organization:
-      - braintrust-data
-      openai-processing-ms:
-      - '3933'
-      openai-project:
-      - proj_vsCSXafhhByzWOThMrJcZiw9
-      openai-version:
-      - '2020-10-01'
-      x-envoy-upstream-service-time:
-      - '3947'
-      x-openai-proxy-wasm:
-      - v0.1
-      x-ratelimit-limit-requests:
-      - '30000'
-      x-ratelimit-limit-tokens:
-      - '150000000'
-      x-ratelimit-remaining-requests:
-      - '29999'
-      x-ratelimit-remaining-tokens:
-      - '149999992'
-      x-ratelimit-reset-requests:
-      - 2ms
-      x-ratelimit-reset-tokens:
-      - 0s
-      x-request-id:
-      - req_cbcce6e1408a47d4b5e8fc61315d4b21
-    status:
-      code: 200
-      message: OK
-version: 1
